# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Real Estate Operations Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List all leads
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: assignedAgentId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Lead list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /leads/{id}/notes:
    get:
      operationId: getLeadNotes
      tags: [leads]
      summary: Get notes for a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createLeadNote
      tags: [leads]
      summary: Add a note to a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /leads/{id}/activity:
    get:
      operationId: getLeadActivity
      tags: [leads]
      summary: Get activity timeline for a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityLog"

  # ─── PROPERTIES ──────────────────────────────────────────────────────────────
  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      summary: List all properties
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Property list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"
    post:
      operationId: createProperty
      tags: [properties]
      summary: Create a property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get a property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Not found
    patch:
      operationId: updateProperty
      tags: [properties]
      summary: Update a property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
    delete:
      operationId: deleteProperty
      tags: [properties]
      summary: Delete a property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CLIENTS (CRM) ───────────────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List all notifications
      parameters:
        - name: read
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notification list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Lead:
      type: object
      required: [id, firstName, lastName, email, status, score, source, createdAt]
      properties:
        id: { type: integer }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        status: { type: string }
        score: { type: integer }
        source: { type: string }
        budget: { type: ["number", "null"] }
        propertyType: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        assignedAgentId: { type: ["integer", "null"] }
        assignedAgentName: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    LeadInput:
      type: object
      required: [firstName, lastName, email, source]
      properties:
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        phone: { type: string }
        status: { type: string }
        score: { type: integer }
        source: { type: string }
        budget: { type: number }
        propertyType: { type: string }
        notes: { type: string }
        assignedAgentId: { type: integer }
    LeadUpdate:
      type: object
      properties:
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        phone: { type: string }
        status: { type: string }
        score: { type: integer }
        source: { type: string }
        budget: { type: number }
        propertyType: { type: string }
        notes: { type: string }
        assignedAgentId: { type: ["integer", "null"] }
...
    Property:
      type: object
      required: [id, title, address, city, price, type, status, bedrooms, bathrooms, area, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        address: { type: string }
        city: { type: string }
        price: { type: number }
        type: { type: string }
        status: { type: string }
        bedrooms: { type: integer }
        bathrooms: { type: integer }
        area: { type: number }
        description: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        assignedAgentId: { type: ["integer", "null"] }
        assignedAgentName: { type: ["string", "null"] }
        createdAt: { type: string }
    PropertyInput:
      type: object
      required: [title, address, city, price, type, bedrooms, bathrooms, area]
      properties:
        title: { type: string }
        address: { type: string }
        city: { type: string }
        price: { type: number }
        type: { type: string }
        status: { type: string }
        bedrooms: { type: integer }
        bathrooms: { type: integer }
        area: { type: number }
        description: { type: string }
        imageUrl: { type: string }
        assignedAgentId: { type: integer }
    PropertyUpdate:
      type: object
      properties:
        title: { type: string }
        address: { type: string }
        city: { type: string }
        price: { type: number }
        type: { type: string }
        status: { type: string }
        bedrooms: { type: integer }
        bathrooms: { type: integer }
        area: { type: number }
        description: { type: string }
        imageUrl: { type: string }
        assignedAgentId: { type: ["integer", "null"] }

    # ─── CLIENT ──────────────────────────────────────────────────────────────────
...
    Note:
      type: object
      required: [id, content, entityType, entityId, createdAt]
      properties:
        id: { type: integer }
        content: { type: string }
        entityType: { type: string }
        entityId: { type: integer }
        authorName: { type: ["string", "null"] }
        createdAt: { type: string }
    NoteInput:
      type: object
      required: [content]
      properties:
        content: { type: string }
        authorName: { type: string }

    # ─── TASK ────────────────────────────────────────────────────────────────────
...
    Notification:
      type: object
      required: [id, title, message, type, isRead, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        message: { type: string }
        type: { type: string }
        isRead: { type: boolean }
        entityType: { type: ["string", "null"] }
        entityId: { type: ["integer", "null"] }
        createdAt: { type: string }

    # ─── ACTIVITY LOG ────────────────────────────────────────────────────────────
    ActivityLog:
      type: object
      required: [id, action, entityType, entityId, description, createdAt]
      properties:
        id: { type: integer }
        action: { type: string }
        entityType: { type: string }
        entityId: { type: integer }
        description: { type: string }
        performedBy: { type: ["string", "null"] }
        createdAt: { type: string }