# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Real Estate Operations Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface Property {
  id: number;
  title: string;
  address: string;
  city: string;
  price: number;
  type: string;
  status: string;
  bedrooms: number;
  bathrooms: number;
  area: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  assignedAgentId?: number | null;
  /** @nullable */
  assignedAgentName?: string | null;
  createdAt: string;
}