# lib/api-zod/src/generated/types/propertyUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Real Estate Operations Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface PropertyUpdate {
  title?: string;
  address?: string;
  city?: string;
  price?: number;
  type?: string;
  status?: string;
  bedrooms?: number;
  bathrooms?: number;
  area?: number;
  description?: string;
  imageUrl?: string;
  /** @nullable */
  assignedAgentId?: number | null;
}